NAS-715: db-defined-reports admin gui import
