+++ packages/desktop/app/index.js
@@ -5,7 +5,6 @@ import { persistStore } from 'redux-persist';
 import Root from './Root';
 import './fonts.scss';
 import './react-toastify.scss';
-import './jsoneditor.scss';
 import { API } from './api/singletons';
 import { registerYup } from './utils/errorMessages';
 import { initStore, restoreSession, authFailure, versionIncompatible } from './store';
