+++ packages/desktop/app/views/administration/reports/ReportAdminView.js
@@ -2,8 +2,6 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 import { TopBar } from '../../../components';
 import { TabDisplay } from '../../../components/TabDisplay';
-import { ReportsEditView } from './edit';
-import { ReportsExportView } from './export';
 import { ReportsImportView } from './import';
 
 const OuterContainer = styled.div`
@@ -30,19 +28,9 @@ const REPORT_TABS = {
 };
 
 export const ReportsAdminView = () => {
-  const [currentTab, setCurrentTab] = useState(REPORT_TABS.EDIT);
+  const [currentTab, setCurrentTab] = useState(REPORT_TABS.IMPORT);
 
   const tabs = [
-    {
-      label: 'Edit',
-      key: REPORT_TABS.EDIT,
-      icon: 'fa fa-edit',
-      render: () => (
-        <TabContainer>
-          <ReportsEditView />
-        </TabContainer>
-      ),
-    },
     {
       label: 'Import',
       key: REPORT_TABS.IMPORT,
@@ -50,16 +38,6 @@ export const ReportsAdminView = () => {
       render: () => (
         <TabContainer>
           <ReportsImportView />
-        </TabContainer>
-      ),
-    },
-    {
-      label: 'Export',
-      key: REPORT_TABS.EXPORT,
-      icon: 'fa fa-file-export',
-      render: () => (
-        <TabContainer>
-          <ReportsExportView />
         </TabContainer>
       ),
     },
