+++ packages/desktop/app/views/administration/reports/import.js
@@ -5,6 +5,7 @@ import { toast } from 'react-toastify';
 import { useApi } from '../../../api';
 import { FormGrid, Heading4, OutlinedButton, SelectInput, TextInput } from '../../../components';
 import { FileChooserInput } from '../../../components/Field/FileChooserField';
+import { useAuth } from '../../../contexts/Auth';
 
 const InnerContainer = styled.div`
   padding: 20px;
@@ -19,19 +20,15 @@ const FormContainer = styled(FormGrid)`
   margin-bottom: 30px;
 `;
 
-const parseFeedback = (name, { userFallback, versionMethod, createdDefinition, versionNumber }) => {
-  const versionText =
-    versionMethod === 'create'
-      ? `Created version ${versionNumber}`
-      : `Updated existing version ${versionNumber}`;
-  const definitionText = createdDefinition ? 'newly created definition' : 'existing definition';
-  const userFallbackTest = userFallback ? ', with user fallback as no userId provided' : '';
-  return `${versionText} for ${definitionText} ${name}${userFallbackTest}`;
-};
+const parseFeedback = (name, { method, versionNumber, createdDefinition }) =>
+  `${method === 'create' ? 'Created' : 'Updated'} version ${versionNumber} for ${
+    createdDefinition ? 'newly created' : 'existing'
+  } definition ${name}`;
 
 export const ReportsImportView = () => {
   const api = useApi();
   const queryClient = useQueryClient();
+  const { currentUser } = useAuth();
   const [name, setName] = useState('');
   const [reportId, setReportId] = useState(null);
   const [file, setFile] = useState(null);
@@ -43,11 +40,9 @@ export const ReportsImportView = () => {
     if (reportId) setReportId(null);
     setName(event.target.value);
   };
-
   const handleChangeFile = event => {
     setFile(event.target.value);
   };
-
   const handleSelectReport = ({ target: { value } }) => {
     setName(reportData.find(report => report.id === value).name);
     setReportId(value);
@@ -56,7 +51,11 @@ export const ReportsImportView = () => {
   const handleSubmit = async () => {
     setSubmitting(true);
     try {
-      const feedback = await api.post('admin/reports/import', { file, name });
+      const feedback = await api.post('admin/reports/import', {
+        file,
+        name,
+        userId: currentUser.id,
+      });
       toast.success(parseFeedback(name, feedback));
       queryClient.invalidateQueries(['reportList']);
       queryClient.invalidateQueries(['reportVersions', reportId]);
@@ -97,7 +96,7 @@ export const ReportsImportView = () => {
         value={file}
         filters={[{ name: 'JSON (.json)', extensions: ['json'] }]}
       />
-      <StyledButton onClick={handleSubmit} disabled={submitting || !name || !file}>
+      <StyledButton onClick={handleSubmit} isSubmitting={submitting} disabled={!name || !file}>
         Import
       </StyledButton>
     </InnerContainer>
