+++ packages/desktop/package.json
@@ -188,7 +188,6 @@
     "webpack-merge": "^4.1.3"
   },
   "dependencies": {
-    "ajv": "^7.2.4",
     "@babel/core": "^7.17.12",
     "@babel/plugin-proposal-class-properties": "^7.17.12",
     "@babel/plugin-proposal-export-default-from": "^7.17.12",
@@ -220,8 +219,6 @@
     "history": "^4.7.2",
     "iso-lang-codes": "^2.1.1",
     "jquery": "^3.3.1",
-    "jsoneditor": "^9.10.0",
-    "jsoneditor-react": "3.1.2",
     "lodash": "^4.17.10",
     "mathjs": "^9.3.0",
     "mime-types": "^2.1.35",
