+++ packages/sync-server/__tests__/admin/reports/data/invalid.json
@@ -1,0 +1,20 @@
+{
+    "versionNumber": 1,
+    "query": "select * from p"atients limit 0",
+    "queryOptions": {
+      "defaultDateRange": "allTime",
+      "dataSources": [
+        "thisFacility"
+      ],
+      "parameters": [
+        {
+          "parameterField": "ParameterSuggesterSelectField",
+          "label": "Area",
+          "name": "locationGroup",
+          "suggesterEndpoint": "locationGroup"
+        }
+      ]
+    },
+    "status": "published",
+    "notes": "Report doing absolutely nothing"
+  }
