+++ packages/sync-server/app/admin/reports/router.js
@@ -1,0 +1,7 @@
+import express from 'express';
+import { getReports, importVersion } from './handlers';
+
+export const reportsRouter = express.Router();
+
+reportsRouter.post('/import', importVersion);
+reportsRouter.get('/', getReports);
