+++ packages/sync-server/config/default.json
@@ -847,9 +847,6 @@
     // NONE is a no-op and does nothing
     "deletionAction": "RENAME"
   },
-  "dbDefinedReports": {
-    "fallbackUser": "admin@tamanu.io"
-  },
   "validateQuestionConfigs": {
     "enabled": false
   },
